+++ src/vue/apps/resume/components/Navbar.vue
@@ -2,7 +2,9 @@
     <header class="blog-header py-3">
         <div class="row flex-nowrap justify-content-between align-items-center">
             <div class="col-4 pt-1">
-                <a class="text-muted" href="#" @click="callFloatMenu($event)">Language {{ getSettedLanguage() }}</a>
+                <a class="text-muted" href="#" @click="callFloatMenu($event)">
+                    <span>Language {{ getSettedLanguage() }}</span>
+                </a>
             </div>
             <div class="col-4 text-center">
                 <a class="blog-header-logo text-dark" href="#">Natanael F. Neto</a>
@@ -35,18 +37,24 @@
                 return available[index].own_escription;
             },
             callFloatMenu(e) {
+                e.stopPropagation()
                 var floatMenu = {
                     isFloatMenu: true,
                     visible: true,
-                    fixed: false, 
-                    xyPosition: {
-                        x:e.clientX,
-                        y:e.clientY
+                    fixed: true,
+                    sourceParams: {
+                        x: e.target.getBoundingClientRect().x,
+                        y: e.target.getBoundingClientRect().y,
+                        width: e.target.getBoundingClientRect().width,
+                        height: e.target.getBoundingClientRect().height,
                     },
-                    xySize: {
+                    floatParams: {
+                        x:e.clientX,
+                        y:e.clientY,
                         width:200,
                         height:200
-                    }
+                    },
+                    options: ['teste','teste2']
                 }
                 this.$emit('updateFloatMenuPosition',floatMenu)
             }